Struct riddle_time::TimeSystem

pub struct TimeSystem { /* fields omitted */ }

The Riddle time system core state.

Manages tracking framerate and timer state.

It is possible to manage the audio system state independantly - the most important thing to note is that ext::TimeSystemExt::process_frame must be called once per “frame”. The default Riddle integration calls that method whenever the riddle::Event::ProcessFrame event is fired.

Implementations

impl TimeSystem

pub fn fps(&self) -> f32

Get the current FPS as calculated based on previous frame durations.

pub fn delta_secs(&self) -> f32

Get the current delta t as calculated based on previous frame durations.

pub fn frame_instant(&self) -> Instant

Get the reference time for this frame. Captured during ext::TimeSystemExt::process_frame.

pub fn register_timer<F>(&self, duration: Duration, callback: F) -> TimerHandle where
F: FnOnce() + Send + 'static,

Register a timer with a callback which will be fired when the time elpases.

The returned handle may be dropped without cancelling the timer.

Example

let val = Arc::new(AtomicBool::new(false));

time_system.register_timer(std::time::Duration::from_millis(200), {
    let val = val.clone();
    move || { val.store(true, Ordering::Relaxed); }
});

// A while later
assert_eq!(true, val.load(Ordering::Relaxed));